requirements.txt
streamlit
pandas
python-dotenv
crewai
langchain
langchain-openai
langchain-community
openai